requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
colorama==0.4.6
Deprecated==1.2.18
deprecation==2.1.0
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-appengine-logging==1.6.0
google-cloud-audit-log==0.3.1
google-cloud-core==2.4.2
google-cloud-logging==3.11.4
google-genai==1.5.0
googleapis-common-protos==1.69.1
grpc-google-iam-v1==0.14.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
opentelemetry-api==1.30.0
packaging==24.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
rsa==4.9
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
websockets==14.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
